requirements.txt
playwright==1.44.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
python-telegram-bot==21.3
asyncio-throttle==1.0.2